[level 2] Title: k진수에서 소수 개수 구하기, Time: 0.18 ms, Memory: 71.4 MB -BaekjoonHub

+++ 프로그래머스/2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -11,7 +11,7 @@ class Solution {
         return answer;
     }
     static public boolean prime(String str){
-        long num = Long.parseLong(str);
+        int num = Integer.parseInt(str);
         int i=0;
         for(i=2;i<=Math.sqrt(num);i++) {
             if (num % i == 0) {
